API: removed the tests/roi.py functions and added all the test to core.tests.roi
